New variant of BC_Title::update for double
Make use of it in CWindowEyedropGUI::update

+++ cinelerra/cwindowtool.C
@@ -464,9 +464,9 @@ CWindowEyedropGUI::CWindowEyedropGUI(MWindow *mwindow, CWindowTool *thread)
 
 void CWindowEyedropGUI::update()
 {
-	red->update((float)mwindow->edl->local_session->red);
-	green->update((float)mwindow->edl->local_session->green);
-	blue->update((float)mwindow->edl->local_session->blue);
+	red->update(mwindow->edl->local_session->red);
+	green->update(mwindow->edl->local_session->green);
+	blue->update(mwindow->edl->local_session->blue);
 	int r = round(CLIP(mwindow->edl->local_session->red, 0, 1) * 0xff);
 	int g = round(CLIP(mwindow->edl->local_session->green, 0, 1) * 0xff);
 	int b = round(CLIP(mwindow->edl->local_session->blue, 0, 1) * 0xff);
